naive matmul
